Fix circular reference in compiler
