+++ packages/compiler/src/config/config-loader.ts
@@ -1,9 +1,9 @@
+import { createSourceFile } from "../core/diagnostics.js";
 import { createDiagnostic } from "../core/messages.js";
 import { getDirectoryPath, isPathAbsolute, joinPaths, resolvePath } from "../core/path-utils.js";
 import { createJSONSchemaValidator } from "../core/schema-validator.js";
 import { CompilerHost, Diagnostic, NoTarget, SourceFile } from "../core/types.js";
 import { deepClone, deepFreeze, doIO, omitUndefined } from "../core/util.js";
-import { createSourceFile } from "../index.js";
 import { getLocationInYamlScript } from "../yaml/index.js";
 import { parseYaml } from "../yaml/parser.js";
 import { YamlScript } from "../yaml/types.js";
