+++ packages/compiler/src/core/linter.ts
@@ -1,5 +1,5 @@
 import { DiagnosticCollector, compilerAssert, createDiagnosticCollector } from "./diagnostics.js";
-import { getLocationContext } from "./index.js";
+import { getLocationContext } from "./helpers/index.js";
 import { createDiagnostic } from "./messages.js";
 import { Program } from "./program.js";
 import { EventEmitter, mapEventEmitterToNodeListener, navigateProgram } from "./semantic-walker.js";
