+++ packages/compiler/src/core/param-message.ts
@@ -1,0 +1,20 @@
+import type { CallableMessage } from "./types.js";
+
+export function paramMessage<T extends string[]>(
+  strings: readonly string[],
+  ...keys: T
+): CallableMessage<T> {
+  const template = (dict: Record<T[number], string>) => {
+    const result = [strings[0]];
+    keys.forEach((key, i) => {
+      const value = (dict as any)[key];
+      if (value !== undefined) {
+        result.push(value);
+      }
+      result.push(strings[i + 1]);
+    });
+    return result.join("");
+  };
+  template.keys = keys;
+  return template;
+}
