+++ packages/compiler/src/index.ts
@@ -1,9 +1,12 @@
+import * as coreIndex from "./core/index.js";
+console.log("Srcindex", coreIndex);
+
 export { ResolveCompilerOptionsOptions, resolveCompilerOptions } from "./config/index.js";
 export * from "./core/index.js";
 export * from "./lib/decorators.js";
 export * from "./server/index.js";
-import * as formatter from "./formatter/index.js";
 export const TypeSpecPrettierPlugin = formatter;
 
+import * as formatter from "./formatter/index.js";
 /** @deprecated Use TypeSpecPrettierPlugin */
 export const CadlPrettierPlugin = TypeSpecPrettierPlugin;
